#51 Move content components from common
