+++ ReportJ.Extension.Chrome/app/js/content/components/popup/popup.scss
